backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-jose[cryptography]==3.3.0
httpx>=0.24.0,<0.25.0
python-dotenv==1.0.0
supabase>=2.3.0,<3.0
pydantic-settings>=2.1.0,<3.0
pytest>=7.4.0,<8.0
pytest-asyncio>=0.23.0,<1.0
python-multipart>=0.0.6,<1.0
